Research Proposal Templates
Here are two research proposal templates that I like to use and a general style guide.   
The pdf is a document that I used when I wrote my first research proposal in my Bachelors and I still reference it today =)
The word doc is one which is used in our chair here at ETH when we write SNF funding applications. 
The ppt has some additional input for scientific writing if you want to skim through that.  

You can see that the structure of the pdf and the docx are relatively similar except that we will not need to have previous work and project schedule from the ETH docx.